Solution from Dictionaries - Lab
4.	Students
You will be receiving names of students, their ID, and a course of programming they have taken in the format "{name}:{ID}:{course}". On the last line, you will receive a name of a course in snake case lowercase letters. You should print only the information of the students who have taken the corresponding course in the format: "{name} - {ID}" on separate lines.